Rename types module to annotations
The standard way to build the FFI bindings in development is to run
`python src/netsnmp/netsnmp_ffi.py`.  Unfortunately, this puts
`src/netsnmp` on `sys.path`, which gets in the way when CFFI tries to
import the stdlib module `types`: It tries to import the incorrect
module.

The best workaround atm seems to be to just rename the internal
module.
